// src/components/catalog-templates/modern-4page/types/ColorCustomization.ts
export interface ColorCustomization {
  textColors: {
    companyName: string
    title: string
    description: string
    productName: string
    productDescription: string
    productPrice: string
    categoryName: string
  }
  backgroundColors: {
    main: string
    cover: string
    productCard: string
    categorySection: string
  }
}